Seperated components and made more dynamic
