--- src/components/match/index.tsx
@@ -1,8 +1,0 @@
-import * as React from "react";
-import Match from "./match";
-
-export default class Points extends React.Component {
-  render() {
-    return <Match />;
-  }
-}
